+++ CHANGES.md
@@ -34,6 +34,7 @@ Changes in 7.0
 - SQL no access to Custom "My Report" program for admin by default in rosariosis.sql
 - JS MarkdownToHTML No MarkDown in text, return raw text in warehouse.js
 - Fix Delete from other Student/User Info tabs in Student.php & User.php
+- Remove deprecated since 4.5 rollover_* action hooks in Rollover.php & Actions.php
 
 Changes in 6.9.4
 ----------------